+++ src/com/galya/business/productioncapacity/model/EconomicActivityClass.java
@@ -2,7 +2,7 @@ package com.galya.business.productioncapacity.model;
 
 public class EconomicActivityClass extends EconomicActivityType {
 
-    public EconomicActivityClass(Integer databaseId, String id, String name, String parentId) {
+    public EconomicActivityClass(Long databaseId, String id, String name, String parentId) {
         super(databaseId, id, name, parentId);
     }
