+++ src/com/galya/business/productioncapacity/model/EconomicActivityDivision.java
@@ -2,7 +2,7 @@ package com.galya.business.productioncapacity.model;
 
 public class EconomicActivityDivision extends EconomicActivityType {
 
-    public EconomicActivityDivision(Integer databaseId, String id, String name, String parentId) {
+    public EconomicActivityDivision(Long databaseId, String id, String name, String parentId) {
         super(databaseId, id, name, parentId);
     }
