+++ src/com/galya/business/productioncapacity/model/EconomicActivityGroup.java
@@ -2,7 +2,7 @@ package com.galya.business.productioncapacity.model;
 
 public class EconomicActivityGroup extends EconomicActivityType {
 
-    public EconomicActivityGroup(Integer databaseId, String id, String name, String parentId) {
+    public EconomicActivityGroup(Long databaseId, String id, String name, String parentId) {
         super(databaseId, id, name, parentId);
     }
