+++ src/com/galya/business/productioncapacity/model/EconomicActivitySection.java
@@ -2,7 +2,7 @@ package com.galya.business.productioncapacity.model;
 
 public class EconomicActivitySection extends EconomicActivityType {
 
-    public EconomicActivitySection(Integer databaseId, String id, String name) {
+    public EconomicActivitySection(Long databaseId, String id, String name) {
         super(databaseId, id, name);
     }
